requirements.txt
dnspython==2.4.2
mongomock==4.1.2
nose==1.3.7
packaging==23.1
PyExecJS==1.5.1
pymongo==4.5.0
python-dotenv==1.0.0
sentinels==1.0.0
six==1.16.0
